UI changes for new API
