+++ bachelor/src/frontend/src/views/MediaSeriesDetailsView.vue
@@ -4,7 +4,7 @@
         class="bg-transparent hover:bg-blue-500 text-blue-700 font-semibold hover:text-white py-2 px-4 border border-blue-500 hover:border-transparent rounded">
         Zurück </button>
         <Suspense>
-        <InventoryDetails />
+        <MediaSeriesDetails />
         <template #fallback>
             <p>loadning...</p>
         </template>
@@ -14,11 +14,9 @@
 
 <script setup>
 
-import { useRouter, useRoute } from 'vue-router';
-import InventoryDetails from '@/components/InventoryDetails.vue';
+import { useRouter } from 'vue-router';
+import MediaSeriesDetails from '../components/MediaSeriesDetails.vue';
 
 const router = useRouter();
-
-
 
 </script>
